Removed Unnecessary If Condition to Filter for Display of all Deliveries
Recreated the "feature/2_elsif" branch and added comments for the purpose of pushing to GitHub.
Previously the changes were made and "git flow feature finish 2_elsif" was entered to merge with
the "integration" branch, but this action also deleted the original "2_elsif" branch that is
required to be submitted to GitHub for evidence of closure of issue raised by Mikel.