Extracted Google Calendar from Users Controller and shifted into Rosters Controller instead.
Users Controller now has less responsibility.

Future action is to replace the 'get', 'post', and 'cancelled' with RESTful replacements.
For instance, instead of the "Ignore" button adding the selected event id into a database column to ignore when getting the feed of Google Calendar events so it is no longer displayed, an alternative would be for the "Ignore" button to divert to the Rosters 'update' action, where the event would be modified on the Google Calendar itself and removed if possible. The "elsif" condition code currently in the "show" action of the Rosters Controller may be used as a starting point to be shifted.